Added Routers for CRUD operations

+++ src/Routes/task.js
@@ -1,11 +1,13 @@
 const express = require("express");
 const router = express.Router();
 const Task = require("../models/Task");
+const auth = require("../middleware/auth");
 
-router.get("/tasks/:id", async (req, res) => {
+//GET TASK BY ID
+router.get("/tasks/:id", auth, async (req, res) => {
   const _id = req.params.id;
   try {
-    const task = await Task.findById(_id);
+    const task = await Task.findOne({ _id, owner: req.user._id });
     if (!task) {
       return res.status(404).send({ error: "Task not found" });
     }
@@ -15,17 +17,19 @@ router.get("/tasks/:id", async (req, res) => {
   }
 });
 
-router.get("/tasks", async (req, res) => {
+//GET ALL TASKS
+router.get("/tasks", auth, async (req, res) => {
   try {
-    const tasks = await Task.find({});
-    res.status(200).send(tasks);
+    await req.user.populate("tasks").execPopulate();
+    res.status(200).send(req.user.tasks);
   } catch (error) {
     res.status(500).send(error);
   }
 });
 
-router.post("/tasks", async (req, res) => {
-  const newTask = new Task(req.body);
+//CREATE TASK
+router.post("/tasks", auth, async (req, res) => {
+  const newTask = new Task({ ...req.body, owner: req.user._id });
   try {
     await newTask.save();
     res.status(201).send(newTask);
@@ -34,7 +38,8 @@ router.post("/tasks", async (req, res) => {
   }
 });
 
-router.patch("/tasks/:id", async (req, res) => {
+//UPDATE TASK
+router.patch("/tasks/:id", auth, async (req, res) => {
   const _id = req.params.id;
   const validUpdates = ["description", "completed"];
   const updates = Object.keys(req.body);
@@ -43,24 +48,39 @@ router.patch("/tasks/:id", async (req, res) => {
     return res.status(400).send({ error: "Invalid Update Perfomred" });
   }
   try {
-    const modifiedTask = await Task.findOneAndUpdate({ _id }, req.body, {
-      new: true,
-      runValidators: true
-    });
-
+    const modifiedTask = await Task.findOne({ _id, owner: req.user._id });
     if (!modifiedTask) {
       return res.status(404).send({ error: "Task not found" });
     }
+
+    updates.forEach(update => (modifiedTask[update] = req.body[update]));
+
+    await modifiedTask.save();
+
     res.status(200).send(modifiedTask);
   } catch (error) {
     res.status(400).send(error);
   }
 });
 
-router.delete("/tasks/:id", async (req, res) => {
+//DELETE ALL TASKS
+router.delete("/tasks", auth, async (req, res) => {
+  try {
+    await Task.deleteMany({ owner: req.user._id });
+    if (!task) {
+      return res.status(404).send({ error: "Tasks not found to deleted" });
+    }
+    res.status(200).send({ message: "All Tasks Deleted" });
+  } catch (error) {
+    res.status(500).send(error);
+  }
+});
+
+//DELETE TASK
+router.delete("/tasks/:id", auth, async (req, res) => {
   const _id = req.params.id;
   try {
-    const task = await Task.findOneAndDelete({ _id });
+    const task = await Task.findOneAndDelete({ _id, owner: req.user._id });
     if (!task) {
       return res.status(404).send({ error: "Task not found to delete" });
     }
